chore: Fix README.txt generator and add tests
- Fixes an issue where the readme.txt file on WordPress' website was
  incorrectly formatted and would render incorrectly on the developer
  page.
